jwt auth on courses and restricts info on roster

+++ api/courses.js
@@ -59,40 +59,68 @@ router.get('/:id', async (req, res, next) => {
 })
 
 router.get('/:id/students', checkJwt, async (req, res, next) => {
-  try {
-  const students = await get_students_by_id(parseInt(req.params.id));
-  if (students) {
-    res.status(200).send(students);
-  } else {
-    next();
+  try{
+    let course = await courses_db.find_by_id(body.courseId);
+    if(course && (jwt.role === 'admin' || (jwt.role === 'instructor' && jwt.sub === course.instructorId))){
+      const students = await get_students_by_id(parseInt(req.params.id));
+      if (students) {
+        res.status(200).send(students);
+      } else {
+        next();
+      }
+    } else {
+        res.status(403).send({"Error": "Unauthorized request"})
+    }
+  } catch (err) {
+    console.error(err);
+    res.status(500).send({
+      error: "Unable to fetch students.  Please try again later."
+    });
   }
-} catch (err) {
-  console.error(err);
-  res.status(500).send({
-    error: "Unable to fetch students.  Please try again later."
-  });
-}
-})
+});
 
 //WARNING: I think this works, but it hasn't been tested, and if you have a better solution, go for it
 router.get('/:id/roster', checkJwt, async (req, res, next) => {
-    const roster = await courses_db.get_students_by_id(req.params.id);
-    var rosterCSV = convertToCSV(roster);
-    res.attachment(`roster_course_${res.params.id}.csv`);
-    res.status(200).send(rosterCSV);
+  let body = req.body
+  let id = body.id;
+  let jwt = req.jwt;
+  try{
+    let course = await courses_db.find_by_id(body.courseId);
+    if(course && (jwt.role === 'admin' || (jwt.role === 'instructor' && jwt.sub === course.instructorId))){
+      const roster = await courses_db.get_students_by_id(req.params.id);
+      var rosterCSV = convertToCSV(roster);
+      res.attachment(`roster_course_${res.params.id}.csv`);
+      res.status(200).send(rosterCSV);
+    } else {
+        res.status(403).send({"Error": "Unauthorized request"})
+    }
+  } catch (err) {
+    console.error(err);
+    res.status(500).send({
+      error: "Error retreving roster.  Please try again later."
+    });
+  }
 })
 
 //==POST==
 router.post('/', checkJwt, async (req, res, next) => {
+  let body = req.body
+  let id = body.id;
+  let jwt = req.jwt;
   if (schemaValidate(req.body, CourseSchema)) {
-    try {
-      const id = await courses_db.create(req.body);
-      res.status(201).send({
-        id: id,
-        links: {
-          course: `/courses/${id}`
-        }
-      });
+    try{
+      let course = await courses_db.find_by_id(body.courseId);
+      if(course && (jwt.role === 'admin' || (jwt.role === 'instructor' && jwt.sub === course.instructorId))){
+        const id = await courses_db.create(req.body);
+        res.status(201).send({
+          id: id,
+          links: {
+            course: `/courses/${id}`
+          }
+        });
+      } else {
+        res.status(403).send({"Error": "Unauthorized request"})
+      }
     } catch (err) {
       console.error(err);
       res.status(500).send({
@@ -107,18 +135,26 @@ router.post('/', checkJwt, async (req, res, next) => {
 });
 
 router.post('/:id/students', checkJwt, async (req, res, next) => {
-  try {
-    if (req.body.add) {
-      req.body.add.forEach(async(item, i) => {
-        const result = await courses_db.enroll_by_id(req.params.id, parseInt(i));
-      });
+  let body = req.body
+  let id = body.id;
+  let jwt = req.jwt;
+  try{
+    let course = await courses_db.find_by_id(body.courseId);
+    if(course && (jwt.role === 'admin' || (jwt.role === 'instructor' && jwt.sub === course.instructorId))){
+      if (req.body.add) {
+        req.body.add.forEach(async(item, i) => {
+          const result = await courses_db.enroll_by_id(req.params.id, parseInt(i));
+        });
+      }
+      if (req.body.remove) {
+        req.body.remove.forEach( async(item, i) => {
+          const result = await courses_db.remove_by_id(req.params.id, parseInt(i));
+        });
+      }
+      res.status(200).send("Success");
+    } else {
+        res.status(403).send({"Error": "Unauthorized request"})
     }
-    if (req.body.remove) {
-      req.body.remove.forEach( async(item, i) => {
-        const result = await courses_db.remove_by_id(req.params.id, parseInt(i));
-      });
-    }
-    res.status(200).send("Success");
   } catch (err) {
     res.status(400).send({
       error: "Request body contains at least one invalid user."
@@ -136,7 +172,6 @@ router.patch('/:id', checkJwt, async (req, res, next) => {
         try{
             let course = await courses_db.find_by_id(body.courseId);
             if(course && (jwt.role === 'admin' || (jwt.role === 'instructor' && jwt.sub === course.instructorId))){
-                //TODO: change assignemnt_db.update_by_id parameters to take a body
                 if(await courses_db.update_by_id(id, body)){
                     res.status(200);
                 }
@@ -160,12 +195,20 @@ router.patch('/:id', checkJwt, async (req, res, next) => {
 
 //==DELETE==
 router.delete('/:id', checkJwt, async (req, res, next) => {
+  let body = req.body
+  let id = body.id;
+  let jwt = req.jwt;
   try {
-    const deleteSuccessful = await courses_db.remove_by_id(parseInt(req.params.id));
-    if (deleteSuccessful) {
-      res.status(204).end();
+    let course = await courses_db.find_by_id(body.courseId);
+    if(course && (jwt.role === 'admin')){
+      const deleteSuccessful = await courses_db.remove_by_id(parseInt(req.params.id));
+      if (deleteSuccessful) {
+        res.status(204).end();
+      } else {
+        next();
+      }
     } else {
-      next();
+        res.status(403).send({"Error": "Unauthorized request"})
     }
   } catch (err) {
     console.error(err);
